// src/model/StateType.ts
export enum StateType {
    account = 'account',
    namespace = 'namespace',
    mosaic = 'mosaic',
    multisig = 'multisig',
    hashLock = 'hashLock',
    secretLock = 'secretLock',
    accountRestriction = 'accountRestriction',
    mosaicRestriction = 'mosaicRestriction',
    metadata = 'metadata',
}